Initial commit, log shipping functionality
This is an initial commit of log shipping, linting, and unit testing.
There is still work to be done here (there are still a couple TODOs
in the unit test), but this can at least be given review and feedback
while the final version is being tested/completed.

Issue: RELENG-2792
Signed-off-by: Eric Ball <eball@linuxfoundation.org>
Change-Id: Id5f822fde87352304f302ae964b5af0fe075bdf9