Update patch set 18
Patch Set 18:

Weird, two similar network resets in a row. This may be because both my tox script and the gradle wrapper are downloading gradle files within seconds of each other. But the first one, which pulls in the bin with curl, is succeeding. It's failing on the second command, which is executed by gradle itself, and is pulling in the "all" package, which is a bit bigger than the zip. But this is the first I've seen this issue, so it's hard to say if it might be a persistent problem, or if there's simply an issue with the server connection right now.

Patch-set: 18