Update patch set 4
Patch Set 4:

> Patch Set 3: Code-Review-1
> 
> There's a race condition on the unstash, update, stash. We need some sort of cross step lock

Done. I also added a "warning" block in the docs to note that it requires the Lockable Resources Plugin. I thought it was part of the Pipeline co-requisites at first, but LF Edge Jenkins (which has a fairly minimal set of plugins) has pipelines without the Lockable Resources Plugin. Or it did when I checked; I went ahead and future-proofed it by adding the plugin now.

Patch-set: 4