Feat: Add Gerrit support to ship logs S3
Handle URL encoding paths for Gerrit. When using Gerrit
with  S3_Bucket the pipeline fails to upload the logs
to the S3 bucket because of URL encoding in the s3 path.

Update the $JOB_NAME to replace the URL encoding in the
change-no path or '/' would be replaced by '%2F' in the
log URL. This also makes it hard for devs to navigate
between changes.'

Ex: AWS S3 fails to create the log path:
vex-yul-odl-jenkins-2/aaa-ci-pipeline/70%2F101670%2F21/25
Now this looks like:
vex-yul-odl-jenkins-2/aa-ci-pipeline_70_101670_21/25/

Re-factor existing code for S3 and Nexus logs server
to keep it simple.

Issue-ID: RELENG-4353
Change-Id: I82b04a10199a1397ebca255c1bc6ff7f5ee04a12
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>